realtime/requirements.txt
django
django-channels
django-bootstrap4
channels-redis


